Changing the flocking so that it allows reloading
Reloading was not working because since the flock at the end of the unicorn upstart script blocked indefinitely it blocked any USR2 signals from getting through to the trap. The loop fixes that by periodically coming out of flock, at which point a USR2 trap can be ran if needed.